+++ RS_React/src/components/Card.tsx
@@ -1,5 +1,5 @@
 import { Link, useSearchParams } from 'react-router-dom';
-import { ISearchResponseItem } from './SearchForm';
+import { ISearchResponseItem } from './SearchForm/SearchForm';
 
 export default function Card(data: ISearchResponseItem) {
   const { name, id, height, weight, sprites } = data;
